[Parts of #7] Wrap & signs messages if specified in config
Currently datasync propagates the unencrypted messages. This messages
might be relayed to a third party. We can't therefore rely on whisper
signature, as the author might not be the same as the relayer.

This commit wraps our current message in a protobuf record with an
optional signature.

If the signature is present it will be used to indicate the author of
the message, otherwise whisper is used.

It's not enabled by default as this is a breaking change and should go
live with v1. Decoding is backward compatible, so old messages will
still be read.

I have also changed the messageID scheme to use:
compressedBytesOfPk + Keccak256 of the encoded unencrypted payload, instead of the whisper ID.